+++ packages/shared/portableExperiences/sagas.ts
@@ -1,11 +1,5 @@
-import { EntityV3Asset, parseUrn, resolveUrlFromUrn } from '@dcl/urn-resolver'
-import { call, takeEvery, debounce, select, put, fork } from 'redux-saga/effects'
-import { trackEvent } from 'shared/analytics'
-import { getFetchContentServer } from 'shared/dao/selectors'
-import { waitForMetaConfigurationInitialization } from 'shared/meta/sagas'
-import { getFeatureFlagVariantValue } from 'shared/meta/selectors'
+import { call, takeEvery, debounce, select, put } from 'redux-saga/effects'
 import { StorePortableExperience } from 'shared/types'
-import { Entity } from 'dcl-catalyst-commons'
 import {
   ADD_DESIRED_PORTABLE_EXPERIENCE,
   REMOVE_DESIRED_PORTABLE_EXPERIENCE
@@ -22,8 +16,7 @@ import {
   UPDATE_ENGINE_PX,
   SHUTDOWN_ALL_PORTABLE_EXPERIENCES,
   ACTIVATE_ALL_PORTABLE_EXPERIENCES,
-  ADD_KERNEL_PX,
-  addKernelPortableExperience
+  ADD_KERNEL_PX
 } from './actions'
 import { getDesiredPortableExperiences } from './selectors'
 
@@ -38,62 +31,6 @@ export function* portableExperienceSaga(): any {
   yield takeEvery(REMOVE_SCENE_PX, handlePortableExperienceChanges)
   yield takeEvery(RELOAD_SCENE_PX, reloadPortableExperienceChanges)
   yield debounce(100 /* ms */, UPDATE_ENGINE_PX, handlePortableExperienceChangesEffect)
-
-  yield fork(fetchInitialPortableExperiences)
-}
-
-function* fetchInitialPortableExperiences() {
-  yield waitForMetaConfigurationInitialization()
-
-  const qs = new URLSearchParams(globalThis.location.search)
-
-  const variants: string[] = qs.has('GLOBAL_PX')
-    ? qs.getAll('GLOBAL_PX')
-    : yield select(getFeatureFlagVariantValue, 'initial_portable_experiences')
-
-  try {
-    if (Array.isArray(variants)) {
-      for (const id of variants) {
-        const parsedUrn: EntityV3Asset = yield parseUrn(id)
-        if (parsedUrn) {
-          const url: string = yield call(resolveUrlFromUrn, id)
-          try {
-            if (url) {
-              const result = yield fetch(url)
-              const json: Entity = yield result.json()
-
-              const mappings = json.content || []
-
-              const px: StorePortableExperience = {
-                id,
-                baseUrl: parsedUrn.baseUrl || (yield select(getFetchContentServer)),
-                mappings,
-                menuBarIcon: json.metadata.menuBarIcon || '',
-                name: json.metadata?.display?.title || 'Unnamed',
-                parentCid: 'main'
-              }
-
-              yield put(addKernelPortableExperience(px))
-            }
-          } catch (err: any) {
-            console.error(err)
-            trackEvent('error', {
-              context: 'fetchInitialPortableExperiences',
-              message: err.message,
-              stack: err.stack
-            })
-          }
-        }
-      }
-    }
-  } catch (err: any) {
-    console.error(err)
-    trackEvent('error', {
-      context: 'fetchInitialPortableExperiences',
-      message: err.message,
-      stack: err.stack
-    })
-  }
 }
 
 // every time the desired portable experiences change, the action `updateEnginePortableExperiences` should be dispatched
